ytaudio v0.1.0 Ytaudio.Ytdata

Functions

detail_playlist(playlist_id)

Fetches details for playlist with id playlist_id.

The playlist is represented as an IO list and has a title, a channel title, a publish date and a description.

detail_video(video_id)

Fetches details for video with id video_id.

The video is represented as an IO list and has a title, a channel title, a publish date and a description.

list_playlist(playlist_id, opts \\ %{})

Fetch a list of videos on a playlist with id playlist_id.

Every video is represented as an IO list and has a title, an id, a channel title and a publish date.

list_playlist_video_ids(playlist_id, opts \\ %{})

Fetch a list of the ids of videos on a playlist with id playlist_id.

recommend(video_id, opts \\ %{})

Fetch a list of recommended videos based on a video with id video_id.

The returned items are videos that are represented as an IO list that contains a title, an id, a channel title and a publish date.

search_by_query(type, query, opts \\ %{})

Fetch a list of the videos or playlists based on type for the search query.

If the returned items are videos they are represented as an IO list that contains a title, an id, a channel title, a publish date and a description.

If the returned items are playlists they are represented as an IO list that contains a title, an id, a channel title and a publish date.